feat: update user model to use Float for weight and height
feat: enhance workout plan schema with DifficultyLevel enum

feat: add optional name field to ConversationBase and update related schemas

feat: improve message schema with reply_to_id and is_edited fields

feat: implement paginated responses for messages and conversations

fix: adjust user profile schema to use Float for weight and height

chore: add pytest as a dependency for unit testing

test: add unit tests for authentication endpoints

test: implement unit tests for health check endpoints

test: create unit tests for meal plan endpoints

test: add unit tests for social endpoints including connections and conversations

test: implement unit tests for user endpoints

test: create unit tests for workout endpoints

docs: add CLAUDE.md for guidelines on code generation

test: add shared test fixtures and setup in conftest.py

+++ app/schemas/user.py
@@ -100,8 +100,8 @@ class UserProfileBase(BaseModel):
     bio: Optional[str] = None
     calorie_goal_daily: Optional[int] = None
     sleep_quality: Optional[str] = None  # poor/fair/good
-    weight: Optional[int] = None  # kg
-    height: Optional[int] = None  # cm
+    weight: Optional[float] = None  # kg
+    height: Optional[float] = None  # cm
 
 
 class UserProfileCreate(UserProfileBase):
@@ -172,7 +172,7 @@ class UserLimitationResponse(UserLimitationBase):
 
 class WeightLossProgressBase(BaseModel):
     """Base weight loss progress schema"""
-    weight: int  # kg
+    weight: float  # kg
 
 
 class WeightLossProgressCreate(WeightLossProgressBase):
